requirements.txt
streamlit
pandas
playwright
gspread
google-auth
